requirements.txt
PyMuPDF==1.24.3
pydantic==2.8.2
Jinja2==3.1.4
Pillow==10.4.0
regex==2024.5.15
python-dateutil==2.9.0.post0
# Optional (for OCR fallback; requires system tesseract)
pytesseract==0.3.10
openai==1.107.3
python-dotenv==1.1.1
